requirements.txt
python-telegram-bot[job-queue]
numpy
matplotlib